Move /lib/utils/prompt to /api/utils/prompt.
