+++ lib/utils/prompt/mock.go
@@ -15,64 +15,15 @@
 package prompt
 
 import (
-	"context"
-	"errors"
-	"sync"
+	"github.com/gravitational/teleport/api/utils/prompt"
 )
 
-type FakeReplyFunc func(context.Context) (string, error)
+// TODO(Joerger): Deprecated in favor of `api/utils/prompt`, delete once references are replaced.
+type (
+	FakeReplyFunc = prompt.FakeReplyFunc
+	FakeReader    = prompt.FakeReader
+)
 
-type FakeReader struct {
-	mu      sync.Mutex
-	replies []FakeReplyFunc
-}
-
-// NewFakeReader returns a fake that can be used in place of a ContextReader.
-// Call Add functions in the desired order to configure responses. Each call
-// represents a read reply, in order.
-func NewFakeReader() *FakeReader {
-	return &FakeReader{}
-}
-
-func (r *FakeReader) IsTerminal() bool {
-	return true
-}
-
-func (r *FakeReader) AddReply(fn FakeReplyFunc) *FakeReader {
-	r.mu.Lock()
-	defer r.mu.Unlock()
-	r.replies = append(r.replies, fn)
-	return r
-}
-
-func (r *FakeReader) AddString(s string) *FakeReader {
-	return r.AddReply(func(context.Context) (string, error) {
-		return s, nil
-	})
-}
-
-func (r *FakeReader) AddError(err error) *FakeReader {
-	return r.AddReply(func(context.Context) (string, error) {
-		return "", err
-	})
-}
-
-func (r *FakeReader) ReadContext(ctx context.Context) ([]byte, error) {
-	r.mu.Lock()
-	if len(r.replies) == 0 {
-		r.mu.Unlock()
-		return nil, errors.New("no fake replies available")
-	}
-
-	// Pop first reply.
-	fn := r.replies[0]
-	r.replies = r.replies[1:]
-	r.mu.Unlock()
-
-	val, err := fn(ctx)
-	return []byte(val), err
-}
-
-func (r *FakeReader) ReadPassword(ctx context.Context) ([]byte, error) {
-	return r.ReadContext(ctx)
-}
+var (
+	NewFakeReader = prompt.NewFakeReader
+)
